claims_adjuster/requirements.txt
streamlit
langchain
langchain-openai
langchain-aws
langchain-google-genai
langchain-community
langgraph
faiss-cpu
pydantic
python-dotenv
tiktoken
